MPISchurComplements.jl

MPISchurComplements.MPISchurComplements — Module

MPISchurComplements

See also https://johnomotani.github.io/MPISchurComplements.jl/.

Solves the matrix system

\[\left(\begin{array}{cc} A & B\\ C & D \end{array}\right) \cdot\left(\begin{array}{c} x\\ y \end{array}\right) =\left(\begin{array}{c} u\\ v \end{array}\right)\]

using a Schur complement factorization. This is useful when the block structure of the matrix can be arranged so that $A$ is significantly simpler to solve than the full matrix.

The package is parallelised using shared- and distributed-memory MPI, details below. It does not provide a solver for the matrix $A$ - this must be supplied at setup.

Schur complement factorization

Use the first row $A.x+B.y=u$ to eliminate $x$ from the second row $C.x+D.y=v$:

\[(D - C.A^{-1}\cdot B).y = v - C.A^{-1}.u.\]

\[S=(D - C.A^{-1}\cdot B)\]

is the 'Schur complement' of $A$. It is a dense matrix, which is computed when the factorization is constructed or updated. Computing the Schur complement matrix, and computing the right-hand-side each time the matrix system is solved both require the application of the action of $A^{-1}$, but otherwise only matrix multiplication and addition/subtraction.

To complete the solve, we also need the action of $S^{-1}$. An LU factorization of $S$ is computed, and used to apply $S^{-1}$. This allows $y$ to be computed

\[y = S^{-1}.(v - C.A^{-1}.u),\]

which can then be substituted back into the first row to find $x$

\[x = A^{-1}.u - A^{-1}.B.y.\]

\[A^{-1}.B\]

is a dense matrix. It must be computed as an intermediate step when constructing $S$, and so can be stored and multiplied as a dense matrix (this is the default). Optionally, when $B$ is sparse and the solver that applies $A^{-1}$ is very fast, it might be more optimal to first calculate $B.y$, and then use the $A^{-1}$ solver for a second time (the first was to calculate $A^{-1}.u$).

Overall there are two 'matrix solves' (to calculate $A^{-1}.u$, and the application of $S^{-1}$), with optionally a second application of $A^{-1}$ as just described. The other operations are matrix multiplications and vector addition/subtraction. Matrix multiplication and vector arithmetic operations are fully parallelised. The $A^{-1}$ operation is external to this package, but is assumed to be well parallelised. The $S^{-1}$ operation may be done in serial using the LU solver from LinearAlgebra, which is likely to be optimal for small $S$ (e.g. roughly smaller than 1000x1000). For larger $S$, a parallelised LU solver is provided in the DenseLUs submodule, which might give some speedup.

Usage

The MPISchurComplement matrix factorization struct is created by calling

mpi_schur_complement(A_factorization, B, C, D, owned_top_vector_entries,
                     owned_bottom_vector_entries; distributed_comm=dcomm)

as a first example using just distributed-MPI parallelism (for shared-memory see below). Other optional arguments are described in the docstring for mpi_schur_complement. dcomm is an MPI communicator including all the ranks that participate in this solve.

The matrix entries, solution vectors and right-hand-side vectors are assumed to be domain-decomposed so that different MPI ranks own different parts of the matrices/vectors. However, matrix/vector entries may be duplicated either on different ranks (which may be useful for halo regions that join different subdomains) or even on the same rank (which might be useful for enforcing periodicity).

owned_top_vector_entries gives the 'global' indices of the locally-owned entries of $u$, and owned_bottom_vector_entries gives the 'global' indices of the locally-owned entries of $v$. owned_top_vector_entries and owned_bottom_vector_entries can contain 'overlapping' indices, that are repeated on the same or different MPI ranks. The complete set of indices in owned_top_vector_entries or owned_bottom_vector_entries across all MPI ranks does not have to be consecutive. For example, this might be useful if the entries in $y$ are some subset of grid points that divides a full grid into disconnected segments - in that case the global indices of the full grid can be used directly for both $x$ and $y$. Internally, the indices are converted to values that increase consecutively from 1, but have the same order as those passed in owned_top_vector_entries and owned_bottom_vector_entries.

A_factorization is an object that applies the action of $A^{-1}$ when called as

ldiv!(x, A_factorization, u)

where x and u are the values at the positions given by owned_top_vector_entries.

By default B, C, and D are assumed to be given at the same points as owned_top_vector_entries and owned_bottom_vector_entries: the rows of B and the columns of C are given by owned_top_vector_entries; the columns of B, the rows of C and both rows and columns of D are given by owned_bottom_vector_entries. Optionally, different ranges can be given for the rows of C or the columns of B and D. [Developer note: it was technically more convenient to allow different rows for C, but logically it would make more sense to allow different columns for C, the same as B and D, as this would allow describing discretisation stencils that use points from different subdomains. This could probably be supported if it was needed?]

The requirements for the input matrices and the right-hand-side and solution vectors at 'overlapping' points are different. The right-hand-side vector should have identical entries for overlapping points on all the processes that own that point; similarly the solution vector will have identical values of everlapping points on all processes that own those points. In contrast, the distributed input matrices should be constructed so that for each overlapping point the sum of the values gathered from all processes that own that point gives the value in the assembled, global matrix.

Shared-memory MPI parallelism

Shared-memory MPI parallism is supported, in addition to distributed memory parallelism. Shared-memory parallelism can increase efficiency as data does not need to be copied between processes that belong to a block of processes that share memory. It requires handling of MPI.Win objects that are associated to the shared-memory arrays - see the 'Complete example' section below.

When shared-memory MPI parallelism is used, the matrices B, C and D, and the right-hand-side b and solution x vectors should all be shared memory arrays. In addition, functions allocate_shared_float and allocate_shared_int must be passed to the mpi_schur_complement() constructor. These will be used to allocate some external buffer arrays, and must be provided because the caller must handle freeing the MPI.Win objects associated with the shared-memory arrays, because MPI.free(win::MPI.Win) is a collective operation, and so cannot be handled by the garbage collector.

Complete example

An example using both distributed- and shared-memory parallelism. Note that FakeMPILU is a place-holder for some efficiently parallelised solver for $A^{-1}$.

using MPISchurComplements
using MPISchurComplements: FakeMPILU
using MPI

n_local_top = 9
n_local_bottom = 3

if !MPI.Initialized()
    MPI.Init()
end

comm = MPI.COMM_WORLD
rank = MPI.Comm_rank(comm)
nproc = MPI.Comm_size(comm)

if nproc > 1 && nproc % 2 != 0
    error("`nproc` not divisible by 2")
end

shared_nproc = min(2, nproc)

distributed_rank = rank ÷ shared_nproc
distributed_nproc = nproc ÷ shared_nproc
shared_comm = MPI.Comm_split(comm, distributed_rank, 0)
shared_comm_rank = MPI.Comm_rank(shared_comm)
distributed_comm = MPI.Comm_split(comm, shared_comm_rank == 0 ? 0 : nothing, 0)

local_win_store_float = MPI.Win[]
allocate_shared_float = (dims...)->begin
    if shared_comm_rank == 0
        dims_local = dims
    else
        dims_local = Tuple(0 for _ ∈ dims)
    end
    win, array_temp = MPI.Win_allocate_shared(Array{Float64}, dims_local,
                                              shared_comm)
    array = MPI.Win_shared_query(Array{Float64}, dims, win; rank=0)
    push!(local_win_store_float, win)
    if shared_comm_rank == 0
        array .= NaN
    end
    MPI.Barrier(shared_comm)
    return array
end

local_win_store_int = MPI.Win[]
allocate_shared_int = (dims...)->begin
    if shared_comm_rank == 0
        dims_local = dims
    else
        dims_local = Tuple(0 for _ ∈ dims)
    end
    win, array_temp = MPI.Win_allocate_shared(Array{Int64}, dims_local,
                                              shared_comm)
    array = MPI.Win_shared_query(Array{Int64}, dims, win; rank=0)
    push!(local_win_store_int, win)
    if shared_comm_rank == 0
        array .= typemin(Int64)
    end
    MPI.Barrier(shared_comm)
    return array
end

A_local = allocate_shared_float(n_local_top, n_local_top)
B_local = allocate_shared_float(n_local_top, n_local_bottom)
C_local = allocate_shared_float(n_local_bottom, n_local_top)
D_local = allocate_shared_float(n_local_bottom, n_local_bottom)
u_local = allocate_shared_float(n_local_top)
v_local = allocate_shared_float(n_local_bottom)
x_local = allocate_shared_float(n_local_top)
y_local = allocate_shared_float(n_local_bottom)

if shared_comm_rank == 0
    A_local .= rand(n_local_top, n_local_top)
    B_local .= rand(n_local_top, n_local_bottom)
    C_local .= rand(n_local_bottom, n_local_top)
    D_local .= rand(n_local_bottom, n_local_bottom)
    u_local .= rand(n_local_top)
    v_local .= rand(n_local_bottom)

    # Ensure overlapping points in u and v are identical on different processes.
    if distributed_rank > 0
        req_left1 = MPI.Isend(@view(u_local[1:1]), distributed_comm; dest=distributed_rank-1)
        req_left2 = MPI.Isend(@view(v_local[1:1]), distributed_comm; dest=distributed_rank-1)
    else
        req_left1 = MPI.REQUEST_NULL
        req_left2 = MPI.REQUEST_NULL
    end
    if distributed_rank < distributed_nproc - 1
        req_right1 = MPI.Irecv!(@view(u_local[end:end]), distributed_comm; source=distributed_rank+1)
        req_right2 = MPI.Irecv!(@view(v_local[end:end]), distributed_comm; source=distributed_rank+1)
    else
        req_right1 = MPI.REQUEST_NULL
        req_right2 = MPI.REQUEST_NULL
    end
    MPI.Waitall([req_left1, req_left2, req_right1, req_right2])
end
MPI.Barrier(shared_comm)

owned_top_vector_entries = distributed_rank*(n_local_top-1)+1:(distributed_rank+1)*(n_local_top-1)+1
owned_bottom_vector_entries = distributed_rank*(n_local_bottom-1)+1:(distributed_rank+1)*(n_local_bottom-1)+1

A_factorization = FakeMPILU(A_local, owned_top_vector_entries,
                            owned_top_vector_entries; comm=distributed_comm,
                            shared_comm)

Alu = mpi_schur_complement(A_factorization, B_local, C_local, D_local,
                           owned_top_vector_entries,
                           owned_bottom_vector_entries; comm, shared_comm,
                           distributed_comm, allocate_shared_float,
                           allocate_shared_int)

ldiv!(x_local, y_local, Alu, u_local, v_local)

# Clean up MPI.Win objects from shared arrays
if local_win_store_float !== nothing
    # Free the MPI.Win objects, because if they are free'd by the garbage collector
    # it may cause an MPI error or hang.
    for w ∈ local_win_store_float
        MPI.free(w)
    end
    resize!(local_win_store_float, 0)
end
if local_win_store_int !== nothing
    # Free the MPI.Win objects, because if they are free'd by the garbage collector
    # it may cause an MPI error or hang.
    for w ∈ local_win_store_int
        MPI.free(w)
    end
    resize!(local_win_store_int, 0)
end

LinearAlgebra.ldiv! — Method

ldiv!(x::AbstractVector, y::AbstractVector, sc::MPISchurComplement,
      u::AbstractVector, v::AbstractVector)

Solve the 2x2 block-structured matrix system

\[\left(\begin{array}{cc} A & B\\ C & D \end{array}\right) \cdot \left(\begin{array}{c} x\\ y \end{array}\right) = \left(\begin{array}{c} u\\ v \end{array}\right)\]

for $x$ and $v$.

Only the locally owned parts of u and v should be passed, and the locally owned parts of the solution will be written into x and y. This means that x and u should correspond to the global indices in sc.owned_top_vector_entries and y and v should correspond to the global indices in sc.owned_bottom_vector_entries. When shared memory is used, x and y must be shared-memory arrays (identical on every process in sc.shared_comm). u and v should also be shared memory arrays (although non-shared identical copies would also work).

MPISchurComplements.find_local_vector_inds — Method

find_local_vector_inds(global_inds, owned_global_inds)

Find indices of global_inds within owned_global_inds. This gives the 'local indices' corresponding to global_inds.

MPISchurComplements.mpi_schur_complement — Method

mpi_schur_complement(A_factorization, B::Union{AbstractMatrix,Nothing,Type},
                     C::Union{AbstractMatrix,Nothing,Type},
                     D::Union{AbstractMatrix,Nothing,Type},
                     owned_top_vector_entries::Union{UnitRange{Int64},Vector{Int64}},
                     owned_bottom_vector_entries::Union{UnitRange{Int64},Vector{Int64}};
                     B_global_column_range::Union{UnitRange{Int64},Vector{Int64},Nothing}=nothing,
                     C_global_row_range::Union{UnitRange{Int64},Vector{Int64},Nothing}=nothing,
                     D_global_column_range::Union{UnitRange{Int64},Vector{Int64},Nothing}=nothing,
                     comm::MPI.Comm=MPI.COMM_WORLD,
                     shared_comm::MPI.Comm=MPI.COMM_SELF,
                     distributed_comm::Union{MPI.Comm,Nothing}=nothing,
                     allocate_shared_float::Union{Function,Nothing}=nothing,
                     allocate_shared_int::Union{Function,Nothing}=nothing,
                     synchronize_shared::Union{Function,Nothing}=nothing,
                     use_sparse=true, separate_Ainv_B=false,
                     parallel_schur=(distributed_comm!==nothing || shared_comm!==nothing),
                     schur_tile_size=nothing, skip_factorization=false,
                     check_lu::Bool=true,
                     timer::Union{TimerOutput,Nothing}=nothing)

Initialise an MPISchurComplement struct representing a 2x2 block-structured matrix

\[\left(\begin{array}{cc} A & B\\ C & D \end{array}\right)\]

A_factorization should be a matrix factorization of A, or an object with similar functionality, that can be passed to the ldiv!() function to solve a matrix system.

B and D are only used to initialize the MPISchurComplement, they are not stored. If sparse=false is passed, a reference to C is stored. Only the locally owned parts of B, C and D should be passed. B, C and D may be modified by this function.

owned_top_vector_entries gives the range of global indices that are owned by this process in the top block of the state vector, owned_bottom_vector_entries the same for the bottom block. The index ranges in owned_top_vector_entries and owned_bottom_vector_entries may have gaps, not being a contiguous range starting from 1 (and similarly for B_global_column_range, C_global_row_range and D_global_column_range, if they are passed, although they should all contain the same set of indices, across all processes, as owned_bottom_vector_entries). This may be useful if the 'bottom block' is actually, for example, a chunk from the middle of a grid, which separates the 'top block' into block-diagonal pieces. The global indexing for the full state vector can be used for both owned_top_vector_entries and owned_bottom_vector_entries, and they will be converted (by shifting indices down to remove the gaps) into a representation internal to the MPISchurComplement object that is contiguous and starts at 1 for each block. In addition, the indices passed do not have to be unique (they can be repeated on different subdomains, and/or within a single subdomain) or monotonically increasing - this allows for periodic dimensions where the final point in the dimension should be identified with the first point. When there are repeated indices, the vector entries passed for the right-hand-side (u and v) are assumed to be identical at the locations given corresponding to repeated indices. Matrix entries in B, C, and D corresponding to each repeated row or column index will be (in effect) summed together into a single entry of an 'assembled' matrix (as an implementation detail, the 'assembled' matrix may not be constructed explicitly, but it serves to define the meaning of repeated matrix entries).

B should contain only the rows corresponding to owned_top_vector_entries and by default the columns corresponding to owned_bottom_vector_entries. If a different range of columns should be included, the corresponding global indices can be passed as B_global_column_range.

C should contain only the columns corresponding to owned_top_vector_entries and by default the rows corresponding to owned_bottom_vector_entries. If a different range of rows should be included, the corresponding global indices can be passed as C_global_row_range.

D should contain only the rows corresponding to owned_bottom_vector_entries and by default the same columns. If a different range of columns should be included, the corresponding global indices passed as D_global_column_range.

When shared-memory MPI parallelism is used, B, C, and D should all be shared memory arrays which are identical on all processes in shared_comm (or non-shared identical copies, but that would be an inefficient use of memory). When owned_top_vector_entries and/or owned_bottom_vector_entries are ranges that overlap between different distributed-MPI ranks, the matrices A, B, C, and D should be set up so that adding together the matrices passed on each distributed-MPI rank gives the full, global matrix (i.e. there should be no double-counting of overlapping entries, but any fraction of the overlap can be passed on any distributed rank, as long as the contributions from each distributed rank add up to the full value).

comm is the MPI communicator containing all processes to be used by MPIDenseLU.

shared_comm is the MPI communicator to use for shared-memory communications.

distributed_comm (if passed) is the MPI communicator for distributed-memory communications. It should include only the ranks that are rank-0 in each shared_comm. By default it will be created from comm, but can be passed in to avoid creating extra MPI.Comms.

allocate_shared_float and allocate_shared_int must be passed when shared_comm is passed. They should be passed functions that will be used to allocate various buffer arrays. allocate_shared_float should return arrays with the same element type as A_factorization, B, C, and D, while allocate_shared_int should return arrays with element type Int64. This is necessary as the MPI shared-memory 'windows' (MPI.Win) have to be stored and (eventually) freed. The 'windows' must be managed externally, as if they are freed by garbage collection, the freeing is not synchronized between different processes, causing errors.

synchronize_shared can be passed a function that synchronizes all processes in shared_comm. This may be useful if a custom synchronization is required for debugging, etc. If it is not passed, MPI.Barrier(shared_comm) will be used.

By default, makes C a sparse matrix. Pass use_sparse=false to force C to be dense.

If B is very sparse, and A_factorization is efficiently parallelised, it might be more efficient to compute A \ (B.y) in two steps (B.y then A \ ()) rather than multiplying by the dense Ainv_dot_B. If separate_Ainv_B=true is passed, this will be done (requires use_sparse=true). There is no saving in setup time because Ainv_dot_B still has to be calculated, to be multiplied by C when calculating schur_complement. There is also no memory saving as a dense B-sized buffer array is needed.

By default, when distributed_comm and/orshared_comm are passed the MPIDenseLUs package is used to factorize/solve the Schur complement matrix (which is dense) using a hybrid distributed+shared-memory MPI parallel implementation (both factorization lu!() and solve ldiv!() phases are parallelised) and when neither communicator is is passed the serial implementation from LinearAlgebra (which uses LAPACK/BLAS) is used. parallel_schur can be passed to force use of the serial (false) or parallel MPIDenseLUs (true) implementations. When using MPIDenseLUs, schur_tile_size can be used to set the tile_size argument to mpi_dense_lu(); the default is to use the smaller of 128 and the largest 2^n smaller than half of the size of the global 'bottom vector'.

skip_factorization=true can be passed to create an MPISchurComplement instance without calculating the factorization corresponding to the input matrices. ldiv!() called with this instance will give incorrect results unless update_schur_complement!() is called first. In this case, B, C, and D can be passed nothing or an element type instead of matrices as the matrix values will not be used.

check_lu=false can be passed to disable checks when performing dense LU factorizations. This may increase the speed of the factorization, but leaves it up to the user to guarantee correctness of the input matrices.

A TimerOutput instance can be passed to timer to record timings of various subroutines.

MPISchurComplements.remove_gaps_in_ranges! — Method

remove_gaps_in_ranges!(distributed_ranges::Vector{Vector{Int64}})

Where distributed_ranges does not include only a contiguous range (possibly overlapping) of indices starting from 1, identify and remove any 'gaps' (indices 1 or greater and less than the largest index in distributed_ranges which are not present in any element of distributed_ranges) from the index ranges.

MPISchurComplements.separate_repeated_indices — Method

separate_repeated_indices(inds)

Find any repeated indices in inds. The first occurence of and index is the 'real' entry, and any subsequent occurences are 'repeats. Returns:

a Vector of 'real' entries

a Vector of the positions within the vector (i.e. the local indices) of the 'real' entries

a 2xn Matrix containing all the repeats, where the two entries in each column are the 'real' index and a corresponding 'repeat'

MPISchurComplements.update_schur_complement! — Method

update_schur_complement!(sc::MPISchurComplement, A, B::AbstractMatrix,
                         C::AbstractMatrix, D::AbstractMatrix)

Update the matrix which is being solved by sc.

A will be passed to lu!(sc.A_factorization, A), so should be as required by the LU implementation being used for sc.A_factorization.

B, C, and D should be the same shapes, and represent the same global index ranges, as the inputs to mpi_schur_complement() used to construct sc. B, C, and D may be modified.

MPISchurComplements.update_sparse_matrix! — Method

update_sparse_matrix!(A::SparseMatrixCSC{Tf,Ti},
                      new_A::SparseMatrixCSC{Tf,Ti}) where {Tf,Ti}

Update the values of A in-place to the values of new_A. May not be ideally efficient because it requires resizing Vectors.

MPISchurComplements.FakeMPILUs — Module

Inefficient distributed-MPI LU factorization, for testing.

Requires local rows of matrix to be passed on each process. Just gathers the matrix and RHS vector to the root process, solves there, and scatters back.

Matrix entries can be repeated (e.g. due to overlaps between different subdomains, or periodic boundary conditions). Repeated entries are added together into a single 'assembled' entry. The corresponding vector positions (in RHS and solution vectors) should also be repeated. The repeated entries in the RHS are ignored. Repeated entries in the solution vector are filled by copying the first appearance.