MPISchurComplements.jl

MPISchurComplements.MPISchurComplements — Module

Use the Schur complement technique to solve a block-2x2 matrix system, parallelised with MPI.

Solve the matrix system

\[\left(\begin{array}{cc} A & B\\ C & D \end{array}\right) \cdot\left(\begin{array}{c} x\\ y \end{array}\right) =\left(\begin{array}{c} u\\ v \end{array}\right)\]

LinearAlgebra.ldiv! — Method

ldiv!(x::AbstractVector, y::AbstractVector, sc::MPISchurComplement,
      u::AbstractVector, v::AbstractVector)

Solve the 2x2 block-structured matrix system

\[\left(\begin{array}{cc} A & B\\ C & D \end{array}\right) \cdot \left(\begin{array}{c} x\\ y \end{array}\right) = \left(\begin{array}{c} u\\ v \end{array}\right)\]

for $x$ and $v$.

Only the locally owned parts of u and v should be passed, and the locally owned parts of the solution will be written into x and y. This means that x and u should correspond to the global indices in sc.owned_top_vector_entries and y and v should correspond to the global indices in sc.owned_bottom_vector_entries. When shared memory is used, x and y must be shared-memory arrays (identical on every process in sc.shared_comm). u and v should also be shared memory arrays (although non-shared identical copies would also work).

MPISchurComplements.find_local_vector_inds — Method

find_local_vector_inds(global_inds, owned_global_inds)

Find indices of global_inds within owned_global_inds. This gives the 'local indices' corresponding to global_inds.

MPISchurComplements.mpi_schur_complement — Method

mpi_schur_complement(A_factorization, B::AbstractMatrix, C::AbstractMatrix,
                     D::AbstractMatrix,
                     owned_top_vector_entries::Union{UnitRange{Int64},Vector{Int64}},
                     owned_bottom_vector_entries::Union{UnitRange{Int64},Vector{Int64}};
                     B_global_column_range::Union{UnitRange{Int64},Vector{Int64},Nothing}=nothing,
                     C_global_row_range::Union{UnitRange{Int64},Vector{Int64},Nothing}=nothing,
                     D_global_column_range::Union{UnitRange{Int64},Vector{Int64},Nothing}=nothing,
                     distributed_comm::MPI.Comm=MPI.COMM_SELF,
                     shared_comm::Union{MPI.Comm,Nothing}=nothing,
                     allocate_array::Union{Function,Nothing}=nothing,
                     synchronize_shared::Union{Function,Nothing}=nothing,
                     use_sparse=true, separate_Ainv_B=false)

Initialise an MPISchurComplement struct representing a 2x2 block-structured matrix

\[\left(\begin{array}{cc} A & B\\ C & D \end{array}\right)\]

A_factorization should be a matrix factorization of A, or an object with similar functionality, that can be passed to the ldiv!() function to solve a matrix system.

B and D are only used to initialize the MPISchurComplement, they are not stored. If sparse=false is passed, a reference to C is stored. Only the locally owned parts of B, C and D should be passed. B, C and D may be modified by this function.

owned_top_vector_entries gives the range of global indices that are owned by this process in the top block of the state vector, owned_bottom_vector_entries the same for the bottom block. The index ranges in owned_top_vector_entries and owned_bottom_vector_entries may have gaps, not being a contiguous range starting from 1 (and similarly for B_global_column_range, C_global_row_range and D_global_column_range, if they are passed, although they should all contain the same set of indices, across all processes, as owned_bottom_vector_entries). This may be useful if the 'bottom block' is actually, for example, a chunk from the middle of a grid, which separates the 'top block' into block-diagonal pieces. The global indexing for the full state vector can be used for both owned_top_vector_entries and owned_bottom_vector_entries, and they will be converted (by shifting indices down to remove the gaps) into a representation internal to the MPISchurComplement object that is contiguous and starts at 1 for each block. In addition, the indices passed do not have to be unique (they can be repeated on different subdomains, and/or within a single subdomain) or monotonically increasing - this allows for periodic dimensions where the final point in the dimension should be identified with the first point. When there are repeated indices, the vector entries passed for the right-hand-side (u and v) are assumed to be identical at the locations given corresponding to repeated indices. Matrix entries in B, C, and D corresponding to each repeated row or column index will be (in effect) summed together into a single entry of an 'assembled' matrix (as an implementation detail, the 'assembled' matrix may not be constructed explicitly, but it serves to define the meaning of repeated matrix entries).

B should contain only the rows corresponding to owned_top_vector_entries and by default the columns corresponding to owned_bottom_vector_entries. If a different range of columns should be included, the corresponding global indices can be passed as B_global_column_range.

C should contain only the columns corresponding to owned_top_vector_entries and by default the rows corresponding to owned_bottom_vector_entries. If a different range of rows should be included, the corresponding global indices can be passed as C_global_row_range.

D should contain only the rows corresponding to owned_bottom_vector_entries and by default the same columns. If a different range of columns should be included, the corresponding global indices passed as D_global_column_range.

When shared-memory MPI parallelism is used, B, C, and D should all be shared memory arrays which are identical on all processes in shared_comm (or non-shared identical copies, but that would be an inefficient use of memory). When owned_top_vector_entries and/or owned_bottom_vector_entries are ranges that overlap between different distributed-MPI ranks, the matrices A, B, C, and D should be set up so that adding together the matrices passed on each distributed-MPI rank gives the full, global matrix (i.e. there should be no double-counting of overlapping entries, but any fraction of the overlap can be passed on any distributed rank, as long as the contributions from each distributed rank add up to the full value).

distributed_comm and shared_comm are the MPI communicators to use for distributed-memory and shared-memory communications.

allocate_array is required when shared_comm is passed. It should be passed a function that will be used to allocate various buffer arrays. It should return arrays with the same element type as A_factorization, B, C, and D. This is necessary as the MPI shared-memory 'windows' (MPI.Win) have to be stored and (eventually) freed. The 'windows' must be managed externally, as if they are freed by garbage collection, the freeing is not synchronized between different processes, causing errors.

synchronize_shared can be passed a function that synchronizes all processes in shared_comm. This may be useful if a custom synchronization is required for debugging, etc. If it is not passed, MPI.Barrier(shared_comm) will be used.

By default, makes C a sparse matrix. Pass use_sparse=false to force C to be dense.

If B is very sparse, and A_factorization is efficiently parallelised, it might be more efficient to compute A \ (B.y) in two steps (B.y then A \ ()) rather than multiplying by the dense Ainv_dot_B. If separate_Ainv_B=true is passed, this will be done (requires use_sparse=true). There is no saving in setup time because Ainv_dot_B still has to be calculated, to be multiplied by C when calculating schur_complement. There is also no memory saving as a dense B-sized buffer array is needed.

MPISchurComplements.remove_gaps_in_ranges! — Method

remove_gaps_in_ranges!(distributed_ranges::Vector{Vector{Int64}})

Where distributed_ranges does not include only a contiguous range (possibly overlapping) of indices starting from 1, identify and remove any 'gaps' (indices 1 or greater and less than the largest index in distributed_ranges which are not present in any element of distributed_ranges) from the index ranges.

MPISchurComplements.separate_repeated_indices — Method

separate_repeated_indices(inds)

Find any repeated indices in inds. The first occurence of and index is the 'real' entry, and any subsequent occurences are 'repeats. Returns:

a Vector of 'real' entries

a Vector of the positions within the vector (i.e. the local indices) of the 'real' entries

a 2xn Matrix containing all the repeats, where the two entries in each column are the 'real' index and a corresponding 'repeat'

MPISchurComplements.update_schur_complement! — Method

update_schur_complement!(sc::MPISchurComplement, A, B::AbstractMatrix,
                         C::AbstractMatrix, D::AbstractMatrix)

Update the matrix which is being solved by sc.

A will be passed to lu!(sc.A_factorization, A), so should be as required by the LU implementation being used for sc.A_factorization.

B, C, and D should be the same shapes, and represent the same global index ranges, as the inputs to mpi_schur_complement() used to construct sc. B, C, and D may be modified.

MPISchurComplements.FakeMPILUs — Module

Inefficient distributed-MPI LU factorization, for testing.

Requires local rows of matrix to be passed on each process. Just gathers the matrix and RHS vector to the root process, solves there, and scatters back.

Matrix entries can be repeated (e.g. due to overlaps between different subdomains, or periodic boundary conditions). Repeated entries are added together into a single 'assembled' entry. The corresponding vector positions (in RHS and solution vectors) should also be repeated. The repeated entries in the RHS are ignored. Repeated entries in the solution vector are filled by copying the first appearance.